Fixed text issue where move preview's right target text was pointing at the wrong mon
-Adjusted int type to u32 to match variable across the code
-Fixed text issue where move preview's right target text was pointing at the wrong mon
-adjusted function to use battler for gAiBattleData rather than battlerPosition

+++ include/battle_controllers.h
@@ -380,7 +380,7 @@ void MoveSelectionCreateCursorAt(u8 cursorPos, u8 arg1);
 void MoveSelectionDestroyCursorAt(u8 cursorPosition);
 void PlayerHandleChooseMove(u32 battler);
 void HandleInputChooseMove(u32 battler);
-void CreateMovePreviewText(u8 BattlerPosition);
+void CreateMovePreviewText(u32 BattlerPosition);
 void HandleInputChooseTarget(u32 battler);
 void HandleInputShowEntireFieldTargets(u32 battler);
 void HandleInputShowTargets(u32 battler);
